Debug release.sh a bit more, it now uploads artifacts.
More progress on this script.  Correctly get the project name and
project owner (which assumes you're running this for a github hosted
project where you are an owner).  Create the release entry and upload
the tarball and detached signature.

What's missing at the moment is posting the body of the release
announcement which is a reformatted git log output.  Next revision.