Drop the GitHub Release entry stuff from release.sh
This part has never really worked well.  The only reason I wanted it
was to attach release artifacts when I make a release, but I'm even
beginning to question whether or not I want to do that.  Right now I
still make the artifacts and do the entry by hand.